refactor: reorganize stopwatch component
