+++ src/App.css
@@ -16,7 +16,7 @@ main {
   font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
 }
 
-.container {
+.stopwatch .container {
   border: 2px solid var(--dark-green);
   border-radius: 10px;
   padding: 2rem;
@@ -26,24 +26,24 @@ main {
   place-items: center;
 }
 
-.container h1 {
+.stopwatch .container h1 {
   font-size: 2rem;
   margin-bottom: 1rem;
 }
 
-.display {
+.stopwatch .display {
   font-family: monospace;
   font-size: 1.5rem;
 }
 
-.buttons {
+.stopwatch .buttons {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   column-gap: 0.5rem;
   margin-top: 1rem;
 }
 
-.buttons button {
+.stopwatch .buttons button {
   padding: 0.5rem;
   border-radius: 5px;
   border: 1px solid var(--dark-green);
@@ -52,24 +52,24 @@ main {
   transition: all 0.1s ease-in-out;
 }
 
-.buttons button:hover {
+.stopwatch .buttons button:hover {
   background-color: var(--green);
 }
 
-.buttons .reset:hover {
+.stopwatch .buttons .reset:hover {
   background-color: var(--red);
 }
 
-.lap-list {
+.stopwatch .lap-list {
   margin-top: 1rem;
   margin-left: 1rem;
 }
 
-.lap-list ul {
+.stopwatch .lap-list ul {
   margin-top: 0.5rem;
 }
 
-.lap-list li {
+.stopwatch .lap-list li {
   margin-left: 1.5rem;
   font-family: monospace;
   font-size: 1rem;
